+++ crucible-cli/CHANGELOG.md
@@ -1,4 +1,4 @@
-# next
+# 0.2 -- 2025-11-09
 
 * Remove `resolveForwardDeclarationsHook`. Instead, use `bindFnHandle` in
   `setupHook`.